simplify the base class
